% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jit.R
\name{jit}
\alias{jit}
\title{Compiles an R function to C++ or Nimble DSL code the first time it is called.}
\usage{
jit(fun, argTypes = list(), returnType = NULL, name = NA,
  backend = "C++")
}
\arguments{
\item{fun}{The function to be compiled.}

\item{argTypes}{An optional list of some or all Numble type specs for function arguments.
If any argument types are missing these will be determined at the first invocation.}

\item{returnType}{An optional Nimble type spec for the return value.
If not provided, this will be determined by calling the uncompiled function once.}

\item{name}{An optional name to be used for the compiled code.}

\item{backend}{The target back-end for compilation. Options include:
'C++', or 'Nimble' for the Nimble DSL.}
}
\description{
The first time the function is called, it will be executed and compiled.
Subsequent calls will run the compiled version.
}
